addressing reviews part 1
